Update ctools-installer.sh
when ctools-installer attempted to download and install the packages, all of the root zip files were stored in a ".tmp/<package>/dist/" folder.  When I first attempted to run the script, it returned errors "no zip found".  When I changed that folder to include the /dist portion, the installation worked.